chore: use playwright (#9345)
* chore: use playwright

* 修改图表

* e2e

* up version
