+++ Dockerfile
@@ -7,7 +7,7 @@ RUN yarn
 
 COPY ./ ./
 
-RUN npm run test:all
+RUN npm run test:e2e
 
 RUN npm run fetch:blocks
