+++ jest.config.js
@@ -1,6 +1,5 @@
 module.exports = {
   testURL: 'http://localhost:8000',
-  testEnvironment: './tests/PuppeteerEnvironment',
   verbose: false,
   extraSetupFiles: ['./tests/setupTests.js'],
   globals: {
